% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.alchemy.R
\name{watson.alchemy.combined}
\alias{watson.alchemy.combined}
\title{WatsonR - Alchemy Language COMBINED Calles}
\usage{
watson.alchemy.combined(utterance)
}
\description{
Alchemy Language Functions -- Receives text and returns MULTPLE Features like keyword, entity, sentiment, tone
 @param String to be Analyzed by Alchemy's combined call
 @return JSON formatted like this http://www.alchemyapi.com/api/combined/textc.html
 @export
}

