% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.alchemy.R
\name{watson.alchemy.test}
\alias{watson.alchemy.test}
\title{WatsonR - Alchemy Language Test}
\usage{
watson.alchemy.test(creds)
}
\arguments{
\item{creds}{file name (a file in JSON format)}
}
\value{
NOTHING - just prints the response (hopefully 200 & API response) on screen
}
\description{
Alchemy Language Functions - Light Test
}
