% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.demo.R
\name{watson.demo}
\alias{watson.demo}
\title{WatsonR - Demonstration of Alchemy, Personality Insights, Tone Analyzer and Natural Language Classifier}
\usage{
watson.demo()
}
\description{
Demonstrates WatsonR - Multiple APIs.  Note NLC Natural Language Classifier takes time to complete training
 @param NONE
 @return NOTHING
 @export
}

