% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.keys.R
\name{watson.keys.enter}
\alias{watson.keys.enter}
\title{WatsonR - Enter API Keys Info Manually}
\usage{
watson.keys.enter()
}
\description{
Prompts User to Enter in Keys.
 KEYS.R FILE load with watson.keys.load is alternate method
}
\details{
@param NONE
 @return NOTHING
 @export
}

