% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.keys.R
\name{watson.keys.load}
\alias{watson.keys.load}
\title{WatsonR - Load API Keys from KEYS.R file}
\usage{
watson.keys.load()
}
\value{
NOTHING
}
\description{
This Function will Load your API keys from KEYS.R file if it is in working directory
You can create / edit this file yourself
}
