% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.checkclassifierstatus}
\alias{watson.nlc.checkclassifierstatus}
\title{WatsonR - Natural Language Classifier (NLC) - STATUS Check}
\usage{
watson.nlc.checkclassifierstatus(classifier_id)
}
\arguments{
\item{Classifier}{ID}
}
\value{
Classifier Status
}
\description{
NLC3
}
