% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.createnewclassifier}
\alias{watson.nlc.createnewclassifier}
\title{WatsonR - Natural Language Classifier (NLC) - CREATE}
\usage{
watson.nlc.createnewclassifier(file, classifiername)
}
\arguments{
\item{CSV}{File of Ground Truth to Train;  Name of the New Classifier}
}
\value{
NOTHING
}
\description{
NLC1
}
