% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.deleteclassifier}
\alias{watson.nlc.deleteclassifier}
\title{WatsonR - Natural Language Classifier (NLC) - DELETE Classifier}
\usage{
watson.nlc.deleteclassifier(kill_classifier)
}
\description{
NLC5
 @param  Classifier ID
 @return Message that you've killed the classifier (but not the service)
 @export
}

