% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.listallclassifiers}
\alias{watson.nlc.listallclassifiers}
\title{WatsonR - Natural Language Classifier (NLC) - LIST}
\usage{
watson.nlc.listallclassifiers()
}
\description{
NLC2
 @param  NOTHING
 @return LIST OF CLASSIFIERS
 @export
}

