% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.processtext}
\alias{watson.nlc.processtext}
\title{WatsonR - Natural Language Classifier (NLC) - PROCESS TEXT
ACCEPT QUERY & RETURN RESULT: classifier and confidence score from text input}
\usage{
watson.nlc.processtext(creds, classifier_id, query_text)
}
\arguments{
\item{creds}{the name of the credentials file in json format}

\item{classifier_id}{the classifier identification uuid}

\item{query_text}{to Pass to Classifier}
}
\value{
Classifiers, with confidence
}
\description{
WatsonR - Natural Language Classifier (NLC) - PROCESS TEXT
ACCEPT QUERY & RETURN RESULT: classifier and confidence score from text input
}
