% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.processtextreturnclass}
\alias{watson.nlc.processtextreturnclass}
\title{WatsonR - Natural Language Classifier (NLC) - PROCESS TEXT}
\usage{
watson.nlc.processtextreturnclass(classifier_id, query_text)
}
\description{
NLC6
 @param  Text to Process
 @return Classes and % Confidence
 @export
}

