% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.status}
\alias{watson.nlc.status}
\title{WatsonR - Natural Language Classifier (NLC)}
\usage{
watson.nlc.status()
}
\description{
Natural Language Classifier
 @param NONE
 @return NOTHING
 @export
}

