% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.pi.R
\name{watson.pi.analyze}
\alias{watson.pi.analyze}
\title{WatsonR - Personality Insights (PI) - Analyze}
\usage{
watson.pi.analyze(text_passage)
}
\arguments{
\item{Long}{text passage to analyze - has to be more than 100 words, better if more than 2500}
}
\value{
52 Psychometric traits - big 5, needs, values, personalit
}
\description{
Personality Insights
}
