% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.pi.R
\name{watson.pi.demo}
\alias{watson.pi.demo}
\title{WatsonR - Personality Insights (PI) - DEMO}
\usage{
watson.pi.demo()
}
\description{
Personality Insights - Demo using a Ronald Reagan Speech #
 @param Pulls Ronald Reagan speech that is included in the library
 @return 52 Psychometric traits - big 5, needs, values, personalit
 @export
}

