% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.stt.R
\name{watson.stt.process}
\alias{watson.stt.process}
\title{WatsonR - Speech to Text (STT) PROCESS - receives WAV File}
\usage{
watson.stt.process(filename)
}
\arguments{
\item{Receives}{WAV File - rather a pointer to where FILE.WAV is}
}
\value{
Transcript (Vanilla) of WAV file; omits all other data
}
\description{
Speech to Text
}
