% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.stt.R
\name{watson.stt.test}
\alias{watson.stt.test}
\title{WatsonR - Speech to Text (STT) TEST - WAV to Transcript}
\usage{
watson.stt.test()
}
\arguments{
\item{NONE}{- WAV file test is part of package}
}
\value{
Transcript (Vanilla) of WAV file; omits all other data
}
\description{
Speech to Text
}
