% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.tone.R
\name{watson.tone.analyze}
\alias{watson.tone.analyze}
\title{WatsonR - Tone Analyzer}
\usage{
watson.tone.analyze(utterance)
}
\value{
NOTHING
}
\description{
Tone Analysis receives an utterance and returns
}
