% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.tts.R
\name{watson.tts.demovoices}
\alias{watson.tts.demovoices}
\title{WatsonR - Text to Speech (TTS) - Round The World Introductions -}
\usage{
watson.tts.demovoices()
}
\value{
Everyone Talks
}
\description{
Text to Speech - LIst of Voices
}

