% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.tts.R
\name{watson.tts.listvoices}
\alias{watson.tts.listvoices}
\title{WatsonR - Text to Speech (TTS) - List Voices Available From services}
\usage{
watson.tts.listvoices()
}
\description{
Text to Speech - LIst of Voices
 @param NONE
 @return list of voices from the TTS service
 @export
}

