% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.tts.R
\name{watson.tts.process}
\alias{watson.tts.process}
\title{WatsonR - Text to Speech (TTS) - Speechify the Transcript; demand the voice you want}
\usage{
watson.tts.process(transcript, voice_number)
}
\description{
Text to Speech
 @param TEXT to be made into audio; and Index 1-12 of Voice (see watson.tts.listvoices() for list)
 @return NOTHING
 @export
}

