% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.checkclassifierstatus}
\alias{watson.nlc.checkclassifierstatus}
\title{WatsonR - Natural Language Classifier (NLC) - STATUS Check}
\usage{
watson.nlc.checkclassifierstatus(creds, classifier_id)
}
\arguments{
\item{creds}{the name of the credentials file in json format}

\item{Classifier}{ID}
}
\value{
Classifier Status
}
\description{
CHECK CLASSIFIER STATUS
IF YOU GET MESSAGE BELOW - YOU NEED TO WAIT!
Can take 15m for some medium size classifers (longer for larger ones)
     "The classifier instance is in its training phase,
     not yet ready to accept classify requests"
}
