% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.createnewclassifier}
\alias{watson.nlc.createnewclassifier}
\title{WatsonR - Natural Language Classifier (NLC) - CREATE}
\usage{
watson.nlc.createnewclassifier(creds, csvfile, classifiername)
}
\arguments{
\item{creds}{the name of the credentials file in json format}

\item{csvfile}{File of Ground Truth to Train;  Name of the New Classifier}

\item{classifiername}{The name to assign to the new classifier}
}
\value{
NOTHING
}
\description{
NLC1
}
