% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.createnewclassifier}
\alias{watson.nlc.createnewclassifier}
\title{WatsonR - Natural Language Classifier (NLC) - CREATE
CREATE NEW CLASSIFIER - post /v1/classifiers -
Creates a classifier with CSV data ## URL below no "/" after base url}
\usage{
watson.nlc.createnewclassifier(creds, csvfile, classifiername, lang)
}
\arguments{
\item{creds}{the name of the credentials file in json format}

\item{csvfile}{File of Ground Truth to Train;  Name of the New Classifier}

\item{classifiername}{The name to assign to the new classifier}

\item{lang}{ISO standard 639 language code for classifier}
}
\value{
initial status
}
\description{
NLC1
}
