% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.deleteclassifier}
\alias{watson.nlc.deleteclassifier}
\title{WatsonR - Natural Language Classifier (NLC) - DELETE Classifier}
\usage{
watson.nlc.deleteclassifier(kill_classifier)
}
\arguments{
\item{Classifier}{ID}
}
\value{
Message that you've killed the classifier (but not the service)
}
\description{
NLC5
}
