% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.listallclassifiers}
\alias{watson.nlc.listallclassifiers}
\title{WatsonR - Natural Language Classifier (NLC) - LIST}
\usage{
watson.nlc.listallclassifiers()
}
\arguments{
\item{NOTHING}{}
}
\value{
LIST OF CLASSIFIERS
}
\description{
NLC2
}
