% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.processtext}
\alias{watson.nlc.processtext}
\title{WatsonR - Natural Language Classifier (NLC) - PROCESS TEXT}
\usage{
watson.nlc.processtext(classifier_id, query_text)
}
\arguments{
\item{Text}{to Pass to Classifier}
}
\value{
Classifiers, with confidence
}
\description{
NLC4
}
