% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.nlc.R
\name{watson.nlc.processtext}
\alias{watson.nlc.processtext}
\title{WatsonR - Natural Language Classifier (NLC) - PROCESS TEXT}
\usage{
watson.nlc.processtext(classifier_id, query_text)
}
\description{
NLC4
 @param  Text to Pass to Classifier
 @return Classifiers, with confidence
 @export
}

