% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.pi.R
\name{watson.pi.demo}
\alias{watson.pi.demo}
\title{WatsonR - Personality Insights (PI) - DEMO}
\usage{
watson.pi.demo(creds)
}
\arguments{
\item{creds}{the name of the credentials file in json format}

\item{Pulls}{Ronald Reagan speech that is included in the library}
}
\value{
52 Psychometric traits - big 5, needs, values, personalit
}
\description{
Personality Insights - Demo using a Ronald Reagan Speech #
}
