% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.stt.R
\name{watson.stt.process}
\alias{watson.stt.process}
\title{WatsonR - Speech to Text (STT) PROCESS - receives WAV File}
\usage{
watson.stt.process(filename)
}
\description{
Speech to Text
 @param Receives WAV File - rather a pointer to where FILE.WAV is
 @return Transcript (Vanilla) of WAV file; omits all other data
 @export
}

