% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watson.tts.R
\name{watson.tts.listvoices}
\alias{watson.tts.listvoices}
\title{WatsonR - Text to Speech (TTS) - List Voices Available From services}
\usage{
watson.tts.listvoices(creds)
}
\arguments{
\item{creds}{the name of the credentials file in json format}
}
\value{
list of voices from the TTS service
}
\description{
Text to Speech - LIst of Voices
}
